requirements.txt
google-adk>=0.2.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
reportlab>=4.0.0
requests>=2.31.0
